backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.9
openai>=1.54.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
